// src/frontend/src/api/client/models/UpdateUser.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type UpdateUser = {
    dw_playlist_id?: string;
    save_dw_weekly?: boolean;
    save_full_playlist?: boolean;
    filter_dislikes?: boolean;
    send_mail?: boolean;
    email?: (string | '');
    send_time?: (string | '');
    is_premium?: boolean;
    refresh_token?: string;
    save_time?: (string | '');
};